Files in dependent roles cause issues with include_vars
It looks like for some versions of Ansible, roles which use this role
as a dependency may also have their 'vars' and 'tasks' files visible to
an include_vars within this role.

To me this is a bug (and quite a serious one) but to avoid issue, work
around it.